app/requirements.txt
flask
mysql-connector-python
gunicorn
pandas
openpyxl
xlrd